Updated Logic & Added Test For API
Implemented new handling for TMDb ID inputs to allow for spaces, which are then trimmed to be read as integers. The script will loop until valid input is provided. Additionally, APIs are now tested before the script runs, and if any API fails, the script will terminate. Requests are configured to be sent every second to maintain smooth operation and avoid hitting rate limits.